+++ packages/storybook/.storybook/manager.js
@@ -1,6 +1,10 @@
 import { addons } from '@storybook/addons';
 import { themes } from '@storybook/theming';
+import './titleAddon';
 
 addons.setConfig({
-  theme: themes.dark,
+  theme: {
+    ...themes.dark,
+    brandImage: "./logo.png"
+  },
 });