+++ packages/storybook/stories/constant/theme/highlight.ts
@@ -4,7 +4,7 @@ export const storybookTheme = HighlightCustomTheme.extends(themes.oneDark, {
   numbersBorderColor: "#1ea7fd",
   backgroundColor: "#272727",
   scrollbarTrackColor: "#242323",
-  srollbarThumbColor: "#1ea7fd",
+  scrollbarThumbColor: "#1ea7fd",
   numbersBackgroundColor: "#242323",
   numbersColor: "#cfcfcf"
 });