+++ packages/highlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lmarcel/highlight
 
+## 2.4.0
+
+### Minor Changes
+
+- Text review and new metadata in storybook
+
 ## 2.3.1
 
 ### Patch Changes